android: add improved library project w/o CameraX
This library attempts to make the C++ fork of the ZXing barcode scanning
library available for Android in a simpler, more flexible way than the
original wrapper.

It tries to be a full replacement of the original Java implementation and
provides (almost) the same result metadata as the orignal Java library.
The generation of barcodes is also supported.

Unlike the official Android wrapper library, it does not include nor
requires CameraX and can be used with the legacy camera API, Camera2 or
CameraX. It can also be used down to API level 9.

This alternative exists because these things are important to me and the
apps I build. I just can't, and doesn't want to use CameraX everywhere
and I don't want to exclude people with old devices.