aar: use CamelCase for Kotlin enums
Because CamelCase is also idiomatic in Kotlin and this allows us
to match these symbols with the C++ side.

Of course, one could argue that this is a library that still can
be used with Java, and so should obey the Java naming conventions
where enums are uppercase underscore-separated to reflect that they
are constants.

But first of all, this is an Android library, and Google made Kotlin
the new default language for Android. So CamelCase enums are a good
fit too, and there's some benefit in matching these symbols with the
C++ side.

More on enum naming conventions in Kotlin:
https://kotlinlang.org/docs/coding-conventions.html#property-names

And the corresponding documentation for Java:
https://docs.oracle.com/javase/tutorial/java/javaOO/enum.html