aar: make all DecodeHints available in wrapper
To gain full control of all native `DecodeHints`.

This should have some overhead because it takes a some time
(around ~500 clocks per seconds in my tests) to decompose the
JVM `DecodeHints` object and to set the native `DecodeHints`
accordingly.

But in reality this doesn't make any noticeable difference.
The total processing time of the top level external function
was still the same. Tested on a Pixel 2, FP2, Moto G1 and
Yotaphone 2.