aar: revert using CamelCase for Kotlin enums
This has too many unreleated and unnecessary consequences.

Sure, from Kotlin's point of view, this would be desirable, because
CamelCase enums would match nicely with the native ones.

On the other hand, using CamelCase makes this library less of a drop-in
replacement for the Java implementation of ZXing. And it would break
other software that uses a scan result from ZXing and expects the old
format name.

So it's probably best to simply stick with the previous upper case
names and not making a fuss about something that would just be nice
to have.

And to avoid inconsistencies, all other new enums for the new
`DecodeHints` object have also been converted to upper case with
underscores.

This reverts commit bac1cc436d302c3a445ee7a4d5f7c85564813d71.