aar: include detected matrix in result
And move the corresponding functionality from `ZXING_EXPERIMENTAL_API`
into production.

Since `ZXING_EXPERIMENTAL_API` includes more than just this extension
(which may have an impact on performance and robustness) and because
`Barcode.symbol()` is returning an `ImageView` which isn't as useful
as the original `BitMatrix` for the Android wrapper.